StoreApps/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
pandas==1.3.5
selenium==4.1.0
requests==2.25.1
websocket-client==0.57.0
webencodings==0.5.1
webdriver-manager==3.5.2
chromedriver==2.24.1
psycopg2-binary==2.9.3
beautifulsoup4==4.10.0
bs4==0.0.1
Unidecode==1.3.2
regex==2020.11.13
pyzmq==22.3.0
psycopg2>=2.5
kafka-python==2.0.2


